refactor(llm): polish-prompt-v2 cleanup (Codex follow-ups)
Address all 7 items Codex flagged on PR #264. No behavior change; pure
code motion, doc fixes, and characterization test additions.

1. Extract shared V2 helpers (V2SystemBase, buildSandwichUserMessage,
   formattingClause) from GeminiPromptBuilder.swift into a new file
   PromptV2Support.swift. OpenAIPromptBuilder no longer reaches across
   files for the sandwich helper.
2. Rewrite stale OpenAIPromptBuilder file-level doc comment. The old
   comment claimed OpenAI shared V2SystemBase and formattingClause;
   OpenAI actually owns its own system text and only uses the shared
   sandwich wrapper. Updated the inline comment at the call site.
3. Document legacyTemplate as a trusted opt-out from V2 anti-instruction
   defense in both builders. Tightened wording per Codex: specifies the
   ${transcript} placeholder path and notes the minimum wrapping that
   still applies.
4. Add opening-tag (<transcript>) delimiter-injection tests to both
   builder test suites. Closing-tag defense was already covered.
5. Add edit-mode characterization tests. Gemini's edit and message modes
   use distinct formatting clauses (edit omits the 3+ threshold);
   OpenAI's are textually identical. Tests are provider-specific per
   Codex correction.

Known gap: case-variant delimiter injection (</TRANSCRIPT>) is not
currently defended. Documented in PromptV2Support.swift; will be
addressed in a follow-up hardening PR.

280 tests pass; release build clean; no dashes.
